fix(performance): remove fs dependency
fixes https://github.com/divyenduz/rusty-image-proxy/issues/4
converts DynamicImage to Buffer in-memory.
